server/pyproject.toml
[project]
name = "astron-claw"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    "python-multipart",
    "sqlalchemy[asyncio]",
    "aiomysql",
    "redis[hiredis]",
    "python-dotenv",
    "alembic",
    "loguru",
    "uvloop",
    "httptools",
    "aiobotocore",
    "aiohttp",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]
